// demos/lib/grid/column.component.d.ts
import { AfterContentInit, ChangeDetectorRef, TemplateRef } from "@angular/core";
import { DataType } from "../data-operations/data-util";
import { IgxGridAPIService } from "./api.service";
import { IgxGridCellComponent } from "./cell.component";
import { IgxCellEditorTemplateDirective, IgxCellFooterTemplateDirective, IgxCellHeaderTemplateDirective, IgxCellTemplateDirective } from "./grid.common";
import { IgxGridComponent } from "./grid.component";
export declare class IgxColumnComponent implements AfterContentInit {
    gridAPI: IgxGridAPIService;
    cdr: ChangeDetectorRef;
    field: string;
    header: string;
    sortable: boolean;
    groupable: boolean;
    editable: boolean;
    filterable: boolean;
    resizable: boolean;
    hasSummary: boolean;
    hidden: boolean;
    movable: boolean;
    width: string;
    maxWidth: string;
    minWidth: string;
    headerClasses: string;
    cellClasses: string;
    index: number;
    formatter: (value: any) => any;
    filteringCondition: (target: any, searchVal: any, ignoreCase?: boolean) => any;
    filteringIgnoreCase: boolean;
    sortingIgnoreCase: boolean;
    dataType: DataType;
    pinned: boolean;
    gridID: string;
    summaries: any;
    readonly defaultMinWidth: string;
    readonly grid: IgxGridComponent;
    bodyTemplate: TemplateRef<any>;
    headerTemplate: TemplateRef<any>;
    footerTemplate: TemplateRef<any>;
    inlineEditorTemplate: TemplateRef<any>;
    readonly cells: IgxGridCellComponent[];
    readonly visibleIndex: number;
    protected _bodyTemplate: TemplateRef<any>;
    protected _headerTemplate: TemplateRef<any>;
    protected _footerTemplate: TemplateRef<any>;
    protected _inlineEditorTemplate: TemplateRef<any>;
    protected _summaries: any;
    protected _hidden: boolean;
    protected _index: number;
    private _defaultMinWidth;
    protected cellTemplate: IgxCellTemplateDirective;
    protected headTemplate: IgxCellHeaderTemplateDirective;
    protected footTemplate: IgxCellFooterTemplateDirective;
    protected editorTemplate: IgxCellEditorTemplateDirective;
    constructor(gridAPI: IgxGridAPIService, cdr: ChangeDetectorRef);
    ngAfterContentInit(): void;
    pin(): boolean;
    unpin(): boolean;
    updateHighlights(oldIndex: number, newIndex: number): void;
    protected check(): void;
}